+++ CleanUp/MainModule/ViewModels/MainViewModel.swift
@@ -69,7 +69,7 @@ final class MainViewModel: MainViewModelProtocol {
         var validAssets = [PHAsset]()
         
         assets.enumerateObjects { asset, _, _ in
-            if asset.creationDate != nil { // Теперь учитываем фото без локации
+            if asset.creationDate != nil {
                 validAssets.append(asset)
             }
         }
