+++ CleanUp/MainModule/Views/MainViewController.swift
@@ -156,32 +156,3 @@ private extension MainViewController {
         return "\(viewModel?.getSelectedCount() ?? 0) selected ⚫️ \(groups.count) groups"
     }
 }
-
-extension MainViewController: UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        groups.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "PhotoGroupCell", for: indexPath) as? PhotoGroupCell else { return UICollectionViewCell() }
-        cell.configure(with: groups[indexPath.section].assets)
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: collectionView.bounds.width, height: 200)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        guard let sectionHeader = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "Header", for: indexPath) as? SectionHeader else { return UICollectionReusableView() }
-        sectionHeader.configure(with: "\(groups[indexPath.section].assets.count) Simillar")
-        return sectionHeader
-    }
-    
-}
-
